QCrop/QCropPackage/requirements.txt
qiskit-nature==0.4
qutip
qiskit-aer
pyscf --prefer-binary
numpy
matplotlib
tqdm